Added dbpedia functionality
dbpedia has structured information from Wikipedia. The description tag
in the response is the first paragraph from a Wikipedia article.

Changed "run" to "getTags"

"getTags" no longer returns several empty arrays.

This can be instead of or in addition to Twitter functionality. It would
be very easy to have this paragraph alongside the displayed tweets.